Simplify framework
There's no need for the framework to define three different kinds of
handlers: web handlers, AJAX handlers and handlers of static files. The
best approach, if one wants to use all the three kinds of handlers in
the same binary, is to let the default mux of http package route the
requests to either a trama mux, a handy mux or a default static server.